style: 주석 달기 & 불필요한 코드 정리
